fix value conversion in OP_PUSH_INT
