+++ vm/src/conversions.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <cstdint>
+
+int64_t ntohl(int64_t net);
