+++ vm/src/vm.cpp
@@ -1,14 +1,7 @@
 #include <cstring>
 #include "interpreter.h"
+#include "conversions.h"
 
-int64_t ntohl(int64_t const net) {
-    int64_t host = 0;
-    for (int i = 0; i < 8; i++) {
-        host <<= 8;
-        host |= (net >> (i * 8)) & 0xFF;
-    }
-    return host;
-}
 
 constant_pool load_constant_pool(const char *bytes, int *ip) {
     // Read the number of strings on a single byte
